Fix the petstore example test with iptables proxy
hypothesis: The old userspace proxier would internally retry connections.  The
new one does not.  When this test comes up, the firewall might not yet be open or
something is causing a long delay and a timeout.  I can't repro this failure
locally, so I am shooting in the dark.  It's sort of plausible.

evidence: I can SSH into the jenkins master that is hung and I can see the hung
curl.  I can run that curl by hand and it works.  I can see that my shell is in
the same netns as that hung curl.