Remove part of subnets validation from strategy
Remove from subnet strategy the check on the two following invariants:

  1) All CIDRs within the same VNI are disjoint

  2) VNIs are contained in a single K8s API namespace

The reason they are removed is that they cannot be reliably enforced
in the subnet strategy code, and will therefore be implemented by
a dedicated controller (which can enforce them reliably) in future
commits. The reason they cannot be enforced in the strategy code is
that in case of concurrent creation of subnets violating the
invariants the violation will only be visible after "some time", but
the strategy code is always executed before subnets are created, i.e,
it cannot wait "some time", whereas a controller can.

Basic validation, such as CIDR format and VNI range, is still performed
in subnet strategy.

Also, does minor refactoring and style change of subnet strategy code.