Some tweaks for performance and visualization
Upped the watch cache size on the extension servers, the default of
100 is too low, changed it to 1000.

Allowed the IPAM controller double the default rate (100/sec) of
apiserver calls.

Now my experiments are rate limited by the network fabric on the
worker nodes.

Tweaked some Grafana panels to be useful with the cadvisor stats from
an IKS cluster.