Merge three separate conflicts list into one
The subnets validator had three separate conflicts lists for efficiency:
one for CIDR-only conflicts, one for namespace-only conflicts, and one
for CIDR-and-namespace conflicts. This commit replaces them with a
single, generic conflict list. This is less efficient but makes for way
simpler code.

Also, this commit refactors the validator code in the light of the fact
that conflicts are no logner characterized.