Switch to usability bool and conflicts condition
This commit replaces the SubnetStatus.ValidationOutcome string field
with the SubnetStatus.Usability bool. Also, in case a subnet is not
usable a SubnetCondition is added to its status to carry information
on the reason why the subnet is not usable.

Also, add a duplicate work suppression cache to the validator to avoid
processing a stale subnet after an update to its status which has not
been reflected in its Informer's cache yet.