Make selection of IPLock consistent with comments
Make the IPAM controller select the newest IPLock when choosing one for
a NetworkAttachment that already owns more than one IPLock. The reason
the newest one is chosen is that it might have been created by an
operator trying to fix things. This is what the comments say, but prior
to this commit the oldest IPLock was picked.