Add optional check on conflicts in subnet strategy
Add check on conflicts in subnet strategy as opposed to having the
check only in the subnets validation controller. The advantage is that
most subnets conflicts will be caught directly by the strategy code
causing subnet creation/update to fail, as opposed to allowing a subnet
to exist and labelling it as not validated like the validation
controller does. Since conflicts are a user's mistake, this commit
makes such mistakes more apparent.

This commit also adds a flag to the API server to optionally disable the
check on conflicts in subnet strategy. The check in the strategy cannot
be reliable: conflicting subnets created at the same time might be
created without noticing the conflicts. For this reason the subnet
validation controller is still needed. But testing such controller when
most conflicts are caught by the subnet strategy code is difficult, hence
the aforementioned flag is added.

This commit also does some refactoring of subnet strategy code.