Add filtering on VNI when validating subnets
This commit affects the network API server code which validates subnets
and the subnets validation controller. Prior to it, when validating a
subnet with VNI X, in both places all the existing subnets were
retrieved and those with a different VNI were manually filtered out.
This commit change is that only subnets with VNI X are retrieved, so
that no manual filtering is needed.