Move panic from controllers to controller manager
Make the controllers managed by the controller manager return an error
in case Informers caches cannot be synced as opposed to panicing, and
make the controller manager code which starts the controllers panic
on the aforementioned errors. This way the controller manager and the
managed controllers are still restarted if the Informers caches cannot
sync, but we get rid of the panics in the controllers code.