Make subnet validator use subnet spec immutability
Make reference enqueuing in subnet update notification handlers
conditional on a UID change. Subnet updates are irrelevant to the
subnet validator: VNI and CIDR, the only information it cares about,
are immutable. The only exception are update notifications triggered
by a delete followed by a re-create. But these entail a UID change.

Also, make additions to conflicts caches conditional to equality
between in-cache owner's UID and live owner UID's. Previously a
check on in-cache owner's VNI and CIDR and live owner's VNI and CIDR
were used. Using UIDs is more straightforward and is made possible
by the fact that VNI and CIDR are allowed to change only when there's
a delete followed by a re-create, which entails a UID change.