Update IPAM after API fields immutability
Replace subnet resource version with subnet UID in anticipation data
used to remember in-flight updates and avoid duplicate work. The
anticipation data should be reset if a subnet's CIDR or VNI changes,
and that's why the resource version was needed. After immutability,
CIDR and VNI can only be changed by deleting and re-creating, and this
entails a UID change. This is more efficient because resource version
also changes on updates which do not change CIDR and VNI.