Update conn-agent after API fields immutability
Remove from the field selector for remote attachments the constraint
that status.hostIP is different than the local hostIP. It was needed
because updates to a local attachment X's spec.node made X transition
from local to remote, while keeping status.hostIP set to the local
hostIP. This could lead to keeping a local interface for a remote
attachment. With immutability, this is no longer needed because if X
is deleted and re-created in a way which makes it transition from local
to remote its status gets reset.

Also, change update notification handlers to enqueue references only if
the update is relevant.

Also, fix a bug in initial interfaces sync that could cause leaks of
interfaces.

Also, revamp some comments and do some minor refactorings.