IPAM controller: fix bugs and optimize it.
Make subnet attachments indexer use namespaced names instead of names as
indexes. This avoids unnecessary work: when a notification about a subnet
X is received  the indexer is used to retrieve all the attachments in X
and enqueue them. Before this tweak, the attachments in subnets with the
same name but different namespaces as X are also enqueued.

Also, make the indexer used to retrieve an attachment's IPLocks use
namespaced names instead of names. This improves efficiency because
otherwise when processing a NetworkAttachment X locks owned by
NetworkAttachments with the same name as X but a different namespace
are returned and processed as well. During processing such locks are
recognized as belonging to NetworkAttachments different than X thanks to
a check on UIDs, but they are classified as timeslippers (e.g. belonging
to a newer/older attachment with the same namespaced name as X) and
misleadingly logged as such even if that's not the case.

Also, fix the code that checks (and retrieves) that the IPLock in the
NetworkAttachment's status has been considered to avoid thrashing of IP
addresses which have been already assigned but whose IPLocks are not yet
in the cache. More precisely, get rid of the `considered` map which
stores the IP addresses locked by the IPLocks in the cache. Such map was
used to check whether the IPLock recorded in the NetworkAttachment status
was in the cache and thus has been considered. This commit makes such
check based on whether lockInStatus has been set as that is less liberal
than relying on `considered` and simpler and more correct. It's simpler
because `lockInStatus` is needed regardless of whether it is used in the
check or not while `considered` is needed only for the check. It's more
correct because what that check is doing is making sure that the IPLock
in a NetworkAttachment's status has been considered --- if that's the
case lockInStatus is set. On the other hand, the fact that the IP address
in the NetworkAttachment's status is in `considered` does not mean that
the IPLock has been considered: the IP might have ended up in
`considered` because of an IPLock on the same IP in a different VNI from
a deleted NetworkAttachment with the same namespaced name as the
NetworkAttachment being processed.

Also, perform less prominent fixes such as variable renames not worth an
explanation here.