Pay some technical debt
Merge Grafana dashboards on worker utilization and worker utilization
spent on calls out because they are related and it's useful to see them
in comparison. Do that for IPAM Controller and Subnet Validator only
because the story is more complex with the Connection Agent and because
it will go through a heavy overhaul soon.

Add comment to subnet strategy code on why existing subnets are checked
to make sure they are well formed. This is far from obvious and needs
explanation.