Make CA always process new local NetworkAttachments
Make the connection agent always process a newly created local
NetworkAttachment. Prior to this commit edge-cases causing a newly
created local NetworkAttachment NOT to be processed are possible
when the connection agent is notified of the new NetworkAttachment
via an update notification. This commit fixes that by adding a check
to the code that decides whether to enqueue in the update notification
handler for local NetworkAttachments.