Empty interfaces sync code
Empty the body of the functions to sync network interfaces at start up.
This is done because future commits will alter code which does not
directly deal with network interfaces sync but that is used by that
code, and I want neither to have to update network interfaces sync code
for any change I make, nor the clutter deriving from compiler errors
and my IDE reporting them.

Syncing of network interfaces will be re-added after the rest of the
code stabilizes.