Make check on attachment being up to date readable
Make check on whether a local NetworkAttachment's status is up to date
more readable, at the cost of adding some useless checks.

More precisely, add checks on whether status.hostIP and status.ifcName
are up to date. They are not needed because there is already a check
on status.macAddress: because of how the status is updated if
status.macAddress is up to date then status.hostIP and status.ifcName
must be as well, and if status.macAddress is out of date it is
guaranteed that both status.ifcName and status.hostIP will be set to
their correct values regardless of whether they are not already.
But this might be non-obvious to a reader so we add the checks to make
the code easier to understand.