Comment out pre-existing network interfaces sync
Comment out pre-existing network interfaces sync because work in
this branch will alter the data structures used by that code even
if the primary goal is not changing that code. This way I don't
have to choose between keeping the errors that would appear in sync
code (and all the noise that my IDE would do about that) and
constantly updating the sync code to avoid that noise.

Sync code will be re-added with the due tweaks once the rest of the
code in the connection agent stabilizes.