Remove enqueue on initial sync of remote interface
Remove enqueue of owning network attachment namespaced name when
syncing a pre-existing remote network interface. The enqueue was
there to prevent network interface leaks. Network interface leaks
were thought to happen based on the assumption that when an informer
is stopped no further notification handlers for that informer are
executed. That assumption is false: stopping an informers stops
ingestion of changes in the cache, but notification handlers for
changes prior to the stop are executed, hence the enqueue is useless.