OvS fabric: fix cookie generation for remote interface
Change the way cookies are computed when creating remote network
interfaces. Cookies are an OvS metadata key-val pair (the key being
"cookie") for arbitrary grouping of OpenFlow flows. The OvS-based
network fabric uses them to pair flows added for the same remote
network interface; this allows to group them and use them to parse
the remote network interface they were added for when all the remote
network interfaces are listed. Both before and after this commit
the cookie is a uint64 computed from the network interface MAC
address, before this commit the computation was not collision-free.
The consequence was the flows added for separate remote network
interfaces could bear the same cookie. This caused errors in the
initial syncing of network interfaces the connection agent does.
This commit makes cookie generation collision-free.