Fix race condition during legacy interfaces sync
Add enqueue of the namespaced name of the owner of a remote network
interface during the initial sync of pre-existing network interfaces.
This is needed because there's no guarantee that the add notification
handler for the owner has executed at the time of the match, neither
as for when it will execute. If that has not happened by the time a
worker processes the deletion of the last local NetworkAttachment in
the owner's virtual network, the layer 1 of the virtual network state
is cleared before the owner's namespaced name could be recorded in it.
This means the owner is not processed, hence its network interface is
not deleted even if it should. Enqueuing the owner's namespaced name
forces processing of the owner, which eliminates the race condition.

Also, add optimization to remote network attachments create
notification handlers: namespaced names are enqeueued only if they
could be added to the layer 1 virtual network state. It the addition
is not done there's no point in processing the network attachment.