Rename "layer" into "stage"
Rename "layer" into "stage" in all variables, types, struct fields,
comments and log messages. The reason is that the affected code
application domain is networking, where the term "layer" has a widely
accepted meaning, but we were not using that term with that meaning,
and this can be really confusing. Abbreviations, such as "l", are
renamed as well (e.g., to "s").