Align subject in the key pair to be the same as the documentation
The https-nginx/Makefile refers to subject as "/CN=nginxsvc/O=nginxsvc"
while the documentation at the "Securing the Service" section
expects it to be "/CN=my-nginx/O=my-nginx"

So if we run using the Makefile instead of the manual process
doumented below, the service creation fails.

This PR aligns the subject in the openssl command with the documentation.

It addresses https://github.com/kubernetes/website/issues/34322

Signed-off-by: Eduardo Patrocinio epatro@gmail.com